+++ packages/core/transpiler/view-generator/src/NodeGenerators/ExpGenerator.ts
@@ -22,39 +22,38 @@ export default class ExpGenerator extends ElementGenerator {
 
     if (props) {
       Object.entries(props).forEach(([key, { value, dependencyIndexArr }]) => {
-        const statement = this.setExpProp(dlNodeName, key, value)
-        if (statement) {
-          this.addInitStatement(statement)
-          dependencyIndexArr = [
-            ...(dependencyIndexArr ?? []),
-            ...(content.dependencyIndexArr ?? []),
-          ]
-          if (dependencyIndexArr.length > 0) {
-            this.addUpdateStatements(dependencyIndexArr, statement)
-          }
+        if (
+          ExpGenerator.lifecycle.includes(
+            key as (typeof ExpGenerator.lifecycle)[number]
+          )
+        ) {
+          return this.addInitStatement(
+            this.addLifecycle(
+              dlNodeName,
+              key as (typeof ExpGenerator.lifecycle)[number],
+              value
+            )
+          )
         }
+        if (key === "element") {
+          this.addInitStatement(this.initElement(dlNodeName, value, true))
+          const updateStatement = this.updateElement(dlNodeName, value, true)
+          if (updateStatement)
+            this.addUpdateStatements(dependencyIndexArr, updateStatement)
+          return
+        }
+        if (key === "onUpdate") {
+          this.addUpdateStatements(
+            content.dependencyIndexArr,
+            this.addOnUpdate(dlNodeName, value)
+          )
+          return
+        }
+        DLError.warn1(key)
       })
     }
 
     return dlNodeName
-  }
-
-  /**
-   * @brief Expression node only supports `element` and `do` props
-   * @param dlNodeName
-   * @param key
-   * @param value
-   * @returns
-   */
-  private setExpProp(
-    dlNodeName: string,
-    key: string,
-    value: t.Expression
-  ): t.Statement {
-    if (key === "element") return this.setElement(dlNodeName, value, true)
-    if (key === "do") return this.addDo(dlNodeName, value)
-
-    return DLError.warn1()
   }
 
   /**